Release 1.03.05
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@454 98333e67-4a24-44d7-a75c-e53540dd3050

+++ branches/release_1_03/xmlrpc-c/include/xmlrpc-c/base.h
@@ -54,7 +54,7 @@ typedef double xmlrpc_double;
 */
 
 #define _XMLRPC_STRUCT_MEMBER_OFFSET(TYPE, MBRNAME) \
-  ((unsigned int)(char*)&((TYPE *)0)->MBRNAME)
+  ((unsigned long)(char*)&((TYPE *)0)->MBRNAME)
 #define _XMLRPC_STRUCT_MEMBER_SIZE(TYPE, MBRNAME) \
   sizeof(((TYPE *)0)->MBRNAME)
 #define XMLRPC_STRUCTSIZE(TYPE, MBRNAME) \
